Move regular_tax_before_credits into alternative_income_tax.py
